pyproject.toml
[project]
name = "people-analytics-app"
version = "0.1.0"
description = "Streamlit People Analytics App for Nestlé"
authors = [
    {name = "Kwadwo Osei"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "snowflake-connector-python (>=3.15.0,<4.0.0)",
    "prophet (>=1.1.6,<2.0.0)",
    "naive (>=0.0.3,<0.0.4)",
    "xgboost (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "plotly-express (>=0.4.1,<0.5.0)",
    "plotly (>=6.1.0,<7.0.0)",
    "pyarrow (<19.0.0)",
    "pycountry (>=24.6.1,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[theme]
primaryColor = "#0A5A9C"
backgroundColor = "#F5F9FF"
secondaryBackgroundColor = "#FFFFFF"
textColor = "#2C3E50"
font = "sans serif"


[layout]
wideMode = true

# .streamlit/secrets.toml

ADZUNA_APP_ID = "your_adzuna_app_id_here"
ADZUNA_APP_KEY = "your_adzuna_api_key_here"

